requirements-common.txt
docker-py==1.8.0
factory-boy==2.7.0
fake-factory==0.5.7
# pycrypto==2.6.1
pytest==3.8.0
pytest-helpers-namespace
pytest-timeout
PyYAML
requests
jinja2
rpdb
six
